requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
slowapi>=0.1.9
